engine: Remove explicit timeline loglevel check
This was probably a premature, and likely misguided, optimization.

The timeline will automatically skip messages when logging is
disabled. The extra check was to avoid some arithmetic in computing
the arguments for the log message. However, that seems to be cheap,
and removing the branch means that both logging- and non-logging-
paths can compile to the same trace mcode (no branch on log level in
Lua code, only in the timeline assembler code.)